pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name            = "edof"
version         = "3.0.0"
description     = "Easy Document Format – programmatic document creation, editing and export"
readme          = "README.md"
license         = { file = "LICENSE" }
requires-python = ">=3.9"
keywords        = ["document", "format", "pdf", "template", "edof", "print"]
classifiers     = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Multimedia :: Graphics",
    "Topic :: Office/Business",
    "Topic :: Software Development :: Libraries :: Python Modules",
]

# ── Core dependencies (always installed) ──────────────────────────────────────
dependencies = [
    "Pillow>=9.1",
]

[project.optional-dependencies]
# pip install edof[pdf]
pdf     = ["reportlab>=4.0"]
# pip install edof[qr]
qr      = ["qrcode[pil]>=7.4"]
# pip install edof[pyqt6]
pyqt6   = ["PyQt6>=6.6"]
# pip install edof[all]
all     = [
    "reportlab>=4.0",
    "qrcode[pil]>=7.4",
    "PyQt6>=6.6",
]
# pip install edof[dev]
dev     = [
    "pytest>=7.0",
    "pytest-cov",
    "black",
    "ruff",
    "mypy",
    "build",
    "twine",
]

[project.urls]
Homepage      = "https://github.com/yourusername/edof"
Documentation = "https://github.com/yourusername/edof#readme"
Repository    = "https://github.com/yourusername/edof"
"Bug Tracker" = "https://github.com/yourusername/edof/issues"
Changelog     = "https://github.com/yourusername/edof/blob/main/CHANGELOG.md"

# ── Tool configuration ────────────────────────────────────────────────────────

[tool.setuptools.packages.find]
where = ["."]
include = ["edof*"]

[tool.setuptools.package-data]
edof = ["py.typed"]

[tool.black]
line-length    = 100
target-version = ["py311"]

[tool.ruff]
line-length = 100
select      = ["E", "F", "W", "I"]

[tool.mypy]
python_version         = "3.9"
strict                 = false
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths    = ["tests"]
addopts      = "-v --tb=short"

[tool.coverage.run]
source = ["edof"]
